Add app store icons, setting screen brightness to 100 after generating qr code, log out confirmation modal, lock screen rotation and add all users virtual keys selection instead of team selection

+++ lib/user_hub.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:virtual_key/admin_teams.dart';
-import 'package:virtual_key/user_teams.dart';
+import 'package:virtual_key/user_keys.dart';
 import 'package:virtual_key/globals.dart';
 
 class UserHub extends StatefulWidget {
@@ -16,19 +16,42 @@ class _UserHubState extends State<UserHub> {
   final storage = const FlutterSecureStorage();
 
   int currentIndex = 1;
-  final List<Widget> children = const [AdminTeams(), UserTeams()];
+  final List<Widget> children = const [AdminTeams(), UserKeys()];
 
   void onTappedBar(int index) {
     setState(() {
       if (index == 2) {
-        isLogged = false;
-        token = '';
-        user = null;
+        showDialog(
+          context: context,
+          builder: (BuildContext context) {
+            return AlertDialog(
+              title: Text('Logout'),
+              content: Text('Are you sure you want to logout?'),
+              actions: <Widget>[
+                TextButton(
+                  child: Text('Cancel'),
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                  },
+                ),
+                TextButton(
+                  child: Text('Logout'),
+                  onPressed: () {
+                    isLogged = false;
+                    token = '';
+                    user = null;
 
-        deleteTokenFromStorage();
-        deleteCachedData();
+                    deleteTokenFromStorage();
+                    deleteCachedData();
 
-        Navigator.pushNamedAndRemoveUntil(context, '/', (_) => false);
+                    Navigator.pushNamedAndRemoveUntil(
+                        context, '/', (_) => false);
+                  },
+                ),
+              ],
+            );
+          },
+        );
       } else {
         currentIndex = index;
       }
@@ -40,10 +63,9 @@ class _UserHubState extends State<UserHub> {
   }
 
   Future<void> deleteCachedData() async {
-      var dir = await getTemporaryDirectory();
-        dir.deleteSync(recursive: true);
+    var dir = await getTemporaryDirectory();
+    dir.deleteSync(recursive: true);
   }
-
 
   @override
   Widget build(BuildContext context) {
